+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `NotificationPermission` return from native SDK no longer raises a casting exception on iOS
 - Resolved infinite loops on logging initialization conditions
+- iOS postprocessing will respect existing entitlement files
 
 ## [3.0.0-beta.3]
 ### Fixed